noise-predictor/requirements.txt
pandas>=2.0.0
numpy>=1.24.0,<2.0.0
scikit-learn==1.3.1
psycopg2-binary>=2.9.0
joblib>=1.3.0